chore(services): move function to services file
